Add missed casting

+++ src/unity.c
@@ -310,7 +310,7 @@ void UnityPrintNumberByStyle(const UNITY_INT number, const UNITY_DISPLAY_STYLE_T
             UNITY_OUTPUT_CHAR('\'');
             if ((number <= 126) && (number >= 32))
             {
-                UNITY_OUTPUT_CHAR(number);
+                UNITY_OUTPUT_CHAR((int)number);
             }
             /* write escaped carriage returns */
             else if (number == 13)
